# Refactor inverse_model (Function 3 of FIM): modularization, speed-up, and validation