--- a/docs/source/conf.py
+++ b/docs/source/conf.py
@@ -3,7 +3,6 @@
 # -- Project information
 from fim import __project__, __version__
 
-
 project = __project__
 copyright = "2023, SSEC-JHU"
 author = "SSEC-JHU"