--- a/fim/legacy/VFM_linearortho_VISS.py
+++ b/fim/legacy/VFM_linearortho_VISS.py
@@ -1,8 +1,9 @@
+from multiprocessing import Pool
+
 import numpy as np
 import pandas as pd
 import scipy
 from scipy.optimize import least_squares, minimize
-from multiprocessing import Pool
 
 depth_indentation = 3.2e-05
 sphere_radius = 5e-4