--- a/fim/refactor/main_VFM.py
+++ b/fim/refactor/main_VFM.py
@@ -105,7 +105,7 @@ def load_hgo_fields(folder):
 if __name__ == "__main__":
     start_time = time.time()
 
-    Model = "linear"  # or 'hgo'
+    Model = "hgo"  # or 'hgo'
 
     if Model.lower() == "linear":
         # === Linear Model ===