--- a/tox.ini
+++ b/tox.ini
@@ -13,8 +13,8 @@ skip_install = true
 deps =
     -r requirements/test.txt
 commands =
-    ruff format . --check {posargs}
-    ruff check . --select E --select F --select I {posargs}
+    ruff format . --check {posargs} --exclude fim/legacy/
+    ruff check . --select E --select F --select I {posargs} --exclude fim/legacy/
     # NB: New projects should consider replacing the above with below!
     # ruff check . {posargs}