… fim/test_data/80um/Ux.npy
… fim/test_data/80um/Uy.npy
… fim/test_data/80um/Uz.npy
… fim/test_data/80um/X.npy
… fim/test_data/80um/Y.npy
… fim/test_data/80um/Z.npy
… fim/test_data/HGO/Ux.npy
… fim/test_data/HGO/Uy.npy
… fim/test_data/HGO/Uz.npy
… fim/test_data/HGO/X.npy
… fim/test_data/HGO/Y.npy
… fim/test_data/HGO/Z.npy
… fim/test_data/HGO/volume_matrix.npy