**Summary:**
1) Created a new refactor/ directory containing cleaner, modular scripts (main_VFM.py, material_model.py, vws_models.py) for inverse modeling (Function 3).
2) Introduced a **MaterialModel** object to encapsulate material properties and handle different model types (e.g., linear, HGO) uniformly, improving code reusability and maintainability.
3) Significantly **optimized** computational performance through vectorization and removal of redundant operations.
4) Added a test_data/ directory for storing sample datasets (80um/, HGO/), used in unit testing and model validation.
5) Moved original scripts (VFM_linearortho_VISS.py, VFM_3D_Indentation_HGO_VISS.py) into a new legacy/ directory to preserve the baseline implementation.

**Restructured Codebase:**

1. Added main_VFM.py as the entry point for optimization using displacement data and model selection.
2. Added material_model.py to encapsulate material-specific logic via a MaterialModel class.
3. Added vws_models.py to host vectorized internal/external virtual work functions and reusable utilities.

**Function Highlights:**

1. calculate_VWS_linear() and calculate_VWS_hgo() now share core logic through a unified compute_virtual_work() utility.
2. The MaterialModel object allows flexible switching between linear and HGO models, improving reusability and maintainability.

**Validation:**
1) Tested the refactored scripts in both linear and HGO modes using the default data set and setting.
2) Output results are **identical** to the legacy implementation.
3) Achieved a >50× **speed-up** compared to the original version

**Original Model Benchmark Results:**
1) _Linear Model_
Estimated parameters: (E1, E2) = [5743.27, 1504.54]
Runtime:
Original version: ~400 seconds
Refactored version: ~8 seconds

2) _HGO Model_
Estimated parameters: (C10, D1, kappa) = [273.74, 8.78e-4, 0.0692]
Runtime:
Original version: ~5000 seconds
Refactored version: ~80 seconds

(_All timings measured on the same 16-core laptop._)